+++ crates/mimic/src/core/traits/mod.rs
@@ -222,6 +222,37 @@ impl_field_value_as!(
 );
 
 ///
+/// Inner
+/// for Newtypes to get the innermost value
+///
+
+pub trait Inner<T> {
+    fn inner(&self) -> &T;
+    fn into_inner(self) -> T;
+}
+
+// impl_primitive_inner
+#[macro_export]
+macro_rules! impl_primitive_inner {
+    ($($type:ty),*) => {
+        $(
+            impl Inner<$type> for $type {
+                fn inner(&self) -> &$type {
+                    &self
+                }
+                fn into_inner(self) -> $type {
+                    self
+                }
+            }
+        )*
+    };
+}
+
+impl_primitive_inner!(
+    bool, f32, f64, i8, i16, i32, i64, i128, String, u8, u16, u32, u64, u128
+);
+
+///
 /// Path
 ///
 /// any node created via a macro has a Path