+++ crates/mimic/src/core/types/nat/nat128.rs
@@ -1,8 +1,8 @@
 use crate::core::{
     Value,
     traits::{
-        FieldValue, NumCast, NumToPrimitive, SanitizeAuto, SanitizeCustom, TypeView, ValidateAuto,
-        ValidateCustom, Visitable,
+        FieldValue, Inner, NumCast, NumToPrimitive, SanitizeAuto, SanitizeCustom, TypeView,
+        ValidateAuto, ValidateCustom, Visitable,
     },
 };
 use candid::CandidType;
@@ -53,6 +53,16 @@ impl FieldValue for Nat128 {
 impl From<u128> for Nat128 {
     fn from(u: u128) -> Self {
         Self(u)
+    }
+}
+
+impl Inner<Self> for Nat128 {
+    fn inner(&self) -> &Self {
+        self
+    }
+
+    fn into_inner(self) -> Self {
+        self
     }
 }