+++ crates/mimic/src/core/types/principal.rs
@@ -2,7 +2,7 @@ use crate::{
     ThisError,
     core::{
         traits::{
-            FieldValue, SanitizeAuto, SanitizeCustom, Storable, TypeView, ValidateAuto,
+            FieldValue, Inner, SanitizeAuto, SanitizeCustom, Storable, TypeView, ValidateAuto,
             ValidateCustom, Visitable,
         },
         value::Value,
@@ -130,6 +130,16 @@ impl FromStr for Principal {
             .map_err(|e| PrincipalError::Wrapped(e.to_string()))?;
 
         Ok(this)
+    }
+}
+
+impl Inner<Self> for Principal {
+    fn inner(&self) -> &Self {
+        self
+    }
+
+    fn into_inner(self) -> Self {
+        self
     }
 }